+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### Added - 30-07-2021
+- Fix array object field parsing
+
 #### Added - 12-06-2021
 - Refactor schema parse into segregation of basic and custom type